+++ src/main/java/io/github/astrapi69/tree/binary/LinkedNode.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.tree;
+package io.github.astrapi69.tree.binary;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -33,7 +33,9 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import lombok.experimental.FieldDefaults;
-import io.github.astrapi69.tree.api.ILinkedNode;
+
+import java.util.LinkedList;
+import java.util.List;
 
 /**
  * The class {@link LinkedNode} is a representation of a binary tree node and can have only one next
@@ -50,14 +52,14 @@ import io.github.astrapi69.tree.api.ILinkedNode;
 @AllArgsConstructor
 @Builder(toBuilder = true)
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class LinkedNode<T> implements ILinkedNode<T>
+public class LinkedNode<T>
 {
 
 	/** The next. */
-	ILinkedNode<T> next;
+	LinkedNode<T> next;
 
 	/** The previous. */
-	ILinkedNode<T> previous;
+	LinkedNode<T> previous;
 
 	/** The value. */
 	T value;
@@ -71,5 +73,99 @@ public class LinkedNode<T> implements ILinkedNode<T>
 	public LinkedNode(T value)
 	{
 		this.value = value;
+	}
+
+
+	/**
+	 * Gets the first {@link LinkedNode} object
+	 *
+	 * @return the first {@link LinkedNode} object
+	 */
+	public LinkedNode<T> getFirst()
+	{
+		LinkedNode<T> first = this;
+		LinkedNode<T> previous = getPrevious();
+		while (previous != null && !previous.isFirst())
+		{
+			previous = previous.getPrevious();
+			first = previous;
+		}
+		return first;
+	}
+
+	/**
+	 * Checks for previous object
+	 *
+	 * @return true, if successful
+	 */
+	public boolean hasPrevious()
+	{
+		return getPrevious() != null;
+	}
+
+	/**
+	 * Checks if this is the first object
+	 *
+	 * @return true, if is first
+	 */
+	public boolean isFirst()
+	{
+		return !hasPrevious();
+	}
+
+	/**
+	 * Gets the next count.
+	 *
+	 * @return the next count
+	 */
+	public int getNextCount()
+	{
+		return getNextLinkedNodes().size();
+	}
+
+	/**
+	 * Gets all next {@link LinkedNode} from the current linked node
+	 *
+	 * @return all next {@link LinkedNode} from the current linked node
+	 */
+	public List<LinkedNode<T>> getNextLinkedNodes()
+	{
+		List<LinkedNode<T>> list = new LinkedList<>();
+		LinkedNode<T> next = this;
+		while (next.hasNext())
+		{
+			next = next.getNext();
+			list.add(next);
+		}
+		return list;
+	}
+
+	/**
+	 * Checks if this tree node has a next object
+	 *
+	 * @return true, if successful
+	 */
+	public boolean hasNext()
+	{
+		return getNext() != null;
+	}
+
+	/**
+	 * Transforms this linked node object to an ordered {@link LinkedList} with all nodes
+	 *
+	 * @return the ordered {@link LinkedList} with all nodes
+	 */
+	public List<LinkedNode<T>> toList()
+	{
+		List<LinkedNode<T>> list = new LinkedList<>();
+		LinkedNode<T> first = getFirst();
+		list.add(first);
+		LinkedNode<T> next = first;
+		while (next.hasNext())
+		{
+			next = next.getNext();
+			list.add(next);
+		}
+		return list;
 	}
 }